refactor: 인증 관련 파일명 수정
