+++ FE/src/pages/AuthCallback/AuthCallback.jsx
@@ -1,7 +1,7 @@
 import { useEffect, useRef } from "react";
 import { useNavigate, useSearchParams } from "react-router-dom";
-import { requestKakaoLogin } from "../../api/AuthApi.js";
-import { useAuthStore } from "../../stores/AuthStore.js";
+import { requestKakaoLogin } from "../../api/kakaoAuthApi.js";
+import { useAuthStore } from "../../stores/useAuthStore.js";
 
 export default function AuthCallback() {
   const navigate = useNavigate();
